+++ src/lib.rs
@@ -824,7 +824,7 @@
 
 #![deny(unused_imports, dead_code, unused_variables)]
 #![recursion_limit = "128"]
-#![doc(html_root_url = "https://docs.rs/juniper-from-schema/0.1.3")]
+#![doc(html_root_url = "https://docs.rs/juniper-from-schema/0.1.4")]
 
 extern crate proc_macro;
 extern crate proc_macro2;